rename dsp_interpolation to interpolation to avoid redundancy
