+++ pyfar/dsp/interpolation.py
